[algorithms.mat_eqn_solvers] reorganize files
